Fill-in association identifiers completely
The provisioning module fill-is in shadowRef to all association values
when returning a shadowed resource object. However, identifiers were
originally not touched - they may or may not be present completely,
depending on the situation. This caused issues where the naming
attribute was expected to evaluate tolerant/intolerant patterns
but only primary identifier (uid) was present.

Now we copy all the identifiers from the resolved shadow, along with
the shadowRef. It should ensure that tolerant/intolerant patterns
can be always evaluated.

Related to MID-8815.