Skip "get" autz application in obvious cases
Here we check for the opportunity to skip cloning and/or traversing
objects during application of "get" authorization.

The idea is that if the whole object is readable, we do not have to
clone it (if immutable) or traverse through all its items and check
the "get" authorization for them.

This reduces the model post-processing times for medium-sized user
object read by #all-privileged user from around 20 ms (in midPoint 4.7)
through 0.25 ms (before this commit) to less than 0.01 ms (now).

Of course, when authorizations have to be applied, the performance is
still at the level of hundreds of microseconds. This may be improved
later, if needed.

No (intentional) behavioral changes should be here. Only optimizations
and documentation.