Reduce object cloning
1. SUCCESS fetch result is no longer recorded into objects returned by
provisioning.getObject operation. (This involved cloning of immutable
objects.)

2. Reference values to be put into roleMembershipRef are no longer
cloned "in full", as it embraced cloning of the embedded object. Simpler
copy operation is invoked instead.

3. "Mark resource up" in resource manager is invoked after successful
get operation only if the resource was previously not marked as "up".
The resource manager checks the current status in not quite optimal way
now, so it's better to avoid that invocation if possible.