MID-8796 Fixed Custom CertificationHandler not working
There was a process in the CertGuiHandlerRegistry class that returned a
CertGuiHandler according to the URI of the CertificationHandler.
This part was previously hard-coded and therefore Custom
CertificationHandler could not be used. This has been fixed so that the
CertGuiHandlerRegistry is automatically registered according to the URL
of the CertificationHandler at start-up.