Remove ApprovalStageExecutionRecordType
This structure was used as a return value by (internal) midPoint
APIs when asking for information about planned or real execution
of an approval case.
(See https://docs.evolveum.com/midpoint/devel/design/how-to-display-approval-case-planned-or-real-execution/)

It was deprecated in 4.1 by 80a957c2ef8e58fdeabf230412aa670db4bec8b0
and is not filled-in since then.

Hence, it can be safely removed. No migration is needed.