Remove deprecated boundaryCharacters
This was a configuration property for bucketing, used primarily
for multi-node tasks. It had problems with unordered nature of prism
items, so it was replaced by sound `boundary` item and deprecated
in 4.3 by 0d0a9dff1ccc62fc39de73e194bb67a4e5322cc5.

The migration consists of converting each `boundaryCharacters` value
to `boundary` item with:

- the `position` property, starting at 1 and increasing,
- the `characters` property, containing original `boundaryCharacters`
  value.