Remove deprecated TaskType.policyRule
There is a migration to focusValidityScan.validityConstraint,
although not a complete one: even if the ability to recompute
objects is retained, custom policy actions can no longer be defined.
However, 1) this functionality was undocumented, untested, and
experimental; 2) since approx. 4.4 it didn't work at all, due to
(unrelated) bug.

Hence, this commit:

1. Fixes the underlying bug. This involves an incompatible change
(improvement) of internal SearchBasedActivityRunSpecifics
interface. This is going to be reflected in the release notes.

2. Removes TaskType.policyRule. If needed, analogous functionality
can be added into FocusValidityScanWorkDefinitionType.