Fix connector schema parsing race condition
The DOM objects used in midPoint are not thread safe, even for reading.
The problem manifests itself when a connector is initialized in
multiple threads at once.

This commit provides an immediate fix by coordinating access to
that particular place. (Note that the issue emerged also due to
- mistaken - schema cloning elimination in
59bee63b1b8eb933db39e8a9b61a4023b25ec4c0, but we are not ready
to re-introduce schema cloning now because of the expected
performance penalty.)

This resolves MID-8860.