Add security to searching/counting containers
The model.countContainers and searchContainers methods got better
security (authorizations) treatment. It is similar to the one in their
object-based counterparts: filter processing before the operation,
and data filtering afterward. (The latter except assignments, as they
currently do not have their owning objects attached.)

In order to do that, here we introduced ValueSelector - a parsed form
of ObjectSelectorType and its subtypes. It is analogous to ObjectFilter
in that it can evaluate itself in a given context. (The context is
currently quite complex, hopefully it will get simpler eventually.)

These selectors reside in the `schema` module. The functionality was
previously in repository service (impl, then api) and in security
enforcer implementation. This may or may not be the final place;
at least some of the clauses may be moved upwards in the future.

Some steps towards more flexible framework for diagnosing selectors
and authorization were taken. However, it's in very early stage now.

Work in progress. The individual parts are still finding their place.
The whole concept of sub-object authorizations in being discussed now.

Other changes:
 - Removed (almost) duplicate WorkItemsAsserter.