Resolve deprecated approvalWorkItems container
OtherPrivilegesLimitationType.approvalWorkItems was deprecated in 4.0,
but it was - by mistake - in use up to (excluding) 4.8. This commit
tries to resolve this: The legacy form is used just as a default value
for the new one (caseManagementWorkItems). This is a preparation for
complete removal of the item in the future.

Other changes:

- Workflow tests are adapted to the recent changes in authorization
processing.