Migrate cert decision recording to new autz
Instead of hardcoded authorization logic for certification decision
recording that used #all and #recordCertificationDecision (for own
decisions), we now use general parameterized #completeWorkItem action.
The legacy #recordCertificationDecision is automatically converted
to the new action on the fly (in memory).

To do that, this commit provides preliminary implementation of
sub-object authorization parameters.