Fix security enforcer logging
As part of improving the authorizations processing, the tracing aspect
was decoupled from the other ones by providing flexible (abstract)
tracing. It is supported now by experimental ProcessingTracer and
AbstractTraceEvent classes.

This commit finalizes this process for security enforcer activities.
All diagnostic logging was fixed and migrated to this new framework.
Although log-based tracing is still the only one supported, the solution
is ready to embrace other options.