Remove ModelInteractionService#canSearch method
This method is currently not used. In the past, it was used to check
whether one is allowed to search role members (to decide if some parts
of GUI should be shown). But we decided to drop this method. So here it
is. Consequently, some other parts of now-unused code could be removed
from SecurityEnforcer.

Other minor simplifications of the enforcer code were done as well.