Audit runAs/runPrivileged properly
The AuditEventRecord was extended by "effectivePrincipalRef"
and "effectivePrivilegesModified" describing the real logged-in
principal carrying out an action, as well as the information whether
its privileges were modified (e.g. by "runPrivileged" directive) or not.

(This commit also changes the creation of MidPointPrincipal objects,
plus other unrelated minor changes.)

Work in progress. The names are preliminary.
The information is not really stored in the audit log yet.