Prepare scripting/library execution profiles
This a step towards execution profiles for scripting (bulk actions)
and function libraries, ultimately aimed at resolution of MID-6913
and MID-7831.

1. Expression profiles (both raw beans and their compiled form) were
enriched with the support for bulk actions and function libraries.

2. ScriptingService interface was cleaned up: legacy methods (dealing
with incomplete data) were removed, background execution was changed
to use ModelInteractionService#submit method, and moved to that service
altogether. Finally, the ScriptingService and underlying scripting
evaluator now expects the execution request to be present as a config
item, with the origin set up.

3. EvaluatedPolicyRule (a potential source of bulk actions) was updated
to expect policy rule config item, instead of "raw" policy rule beans.
Also, getEnabledActions and similar methods now returns config items.
That methods were also "hardened" to fail if the rule was not yet
evaluated (preventing from returning nonsense data).

4. Activity work definition (another source of bulk actions) now has
the information on origin as well.

5. ExpressionProfileManager was factored out from ArchetypeManager.

Work in progress. This commit is mainly in order to merge recent
(conflicting) activity-related changes from other branches.