Check profiles for bulk actions and notifications
1. Scripting service now establishes the expression profile based on
the origin of the script (bulk action) and enforces it.
Hence, the "am I root?" checks for "execute-script",
"evaluate-expression" and "notify" action (having custom notifier)
are now gone. They are replaced by the default profile for bulks,
which is either "full" (for #all autz holder) or "limited" (otherwise).

2. Notifications module now has a rudimentary support for expression
profiles: As for custom notifiers, it requires a caller to provide
explicit expression profile for their execution. (This is used e.g.
from the bulk actions.) Other notifiers run currently under "full"
profile; we may add the profiles for e.g. message template objects
later.

The ad-hoc handler object was removed from Event, and should now be
provided as a parameter to the NotificationManager#processEvent instead.
Internal APIs/SPIs in the module were adapted:
ConfigurationItem, EventProcessingContext and SendingContext objects
are now universally used.

Work in progress:
- Custom case notifications use "full" profile for now.
- Default profiles for bulks/notifications from the system config
are not applied.
- Final review of the use of profiles in bulk actions and custom
notifications should be done.
- Some tests still fail.
- Not everything for MID-6913/MID-7831 is implemented, although we
are drawing nearer.