Fix double evaluation of policy rules
The recently introduced checks for double evaluation of policy rules
revealed that it occurs when a focus is being deleted. This commit fixes
that.

This should fix failing TestLinkedObjects.

Also, introducing highly experimental "logical stack trace" that uses
OperationResult objects to provide a context of an error meaningful
to the engineer.