Add role membership management activity handler
That handler assigns a (presumably) business role to selected members,
while unassigning (presumably) application roles induced by that role.

Besides the handler itself, two features are of interest here:

1. RoleManagementUtil that could encapsulate knowledge about how
exactly we treat e.g. induced roles or matching assignments;

2. New ObjectSetType.objectRef item that allows explicit enumeration
of objects to be processed.

Work in progress. Large sets of objects to be processed are not
supported yet.