Switch expression profile on library function call
When a library function is called, the expression profile is determined
anew for the function; the caller's profile is NOT used. This allows
for calling trusted functions from untrusted (or little-trusted) code.

Also created activity-aware submitTaskFromTemplate method
(in ModelInteractionService) and adapted WebComponentUtil to use it.

Work in progress. Related to MID-6913.