Remove 3rd (extra) audit for accesses metadata
This commit eliminates extra audit record when accesses metadata are
computed for new assignments. The cause of the problem lies in fact that
the accesses metadata for new assignments are not "final", as the IDs
of assignments are not available at the time of their creation. They
are computed only when the user is committed into the repository,
triggering re-computation of the metadata and another cycle of repo
modify + audit.

The solution lies in pre-allocating IDs from the repository, and using
them for both accesses metadata and the final storage of assignments.
This way, only one write and therefore only one execution audit record
is needed.

Work in progress; to be reviewed.

Related to MID-8659.