Add bulk-3#xxx authorizations
The model-3#executeBulkAction (a recent replacement of #executeScript)
was not bad, but even better is providing authorizations for individual
bulk actions: add, delete, enable, disable, ... This way, the admin
is able to fine-tune authorizations to cover exactly what a user
needs to have. Moreover, it is well-aligned to similar namespaces:
gui-3 and rest-3.

Other changes:
- Removed ScriptExecutionException. Bulk actions executor now throws
standard exceptions (SchemaException, ObjectNotFoundException, ...).
- The "search" instruction is now an action; although it cannot
be called dynamically because of a conflict between "type" property
in <action> and in <search>, it is really something that we want
to allow/deny in expression profiles and by authorizations.
- Improved the API by introducing BulkActionExecutionOptions.
- Created BulkAction enum that lists all known actions.