Guarantee expression profiles in mappings
The expression profile in a mapping evaluation is now determined safely
(assuming that the origin of the mapping was determined correctly).

This is assured by:

1. not allowing "undetermined" origins anymore,
2. by treating "external" origin according to principal's authorizations
(just like we do for bulk actions).

Note that the "undetermined unsafe" origin is not used in midPoint
any longer.

Work in progress. Some tests may fail.