Reload data-dependent projection before its wave
If there is a data dependency between projection (like an AD group) and
its dependee (like an OU in which the group resides), and the dependee
changes, we will reload the dependent projection on the start of its
wave.

This resolves MID-8929.

Note it's a preliminary fix, as it e.g. does not discriminate between
relevant and irrelevant changes. Also, maybe it would be sufficient
to load the full shadow ONLY at the beginning of the wave related to
the dependent projection. This can be optimized later. See MID-9083.