Create result for simulation activities by default
It is no longer needed to specify reporting/simulationResult
for simulation activities. The default for whether result is created
is determined from the execution mode. Note that the old-style
simulations for the determination of thresholds should be modified
to turn off the result creation to avoid performance loss.

This should resolve MID-8936.