Limit use of ARCHIVED activation status
This activation status was only "half-supported" from the beginning.
For example, it cannot be mapped onto ConnId __ENABLE__ attribute
(which is a boolean) and it's not supported by the simulated activation
feature either. Gradually, it became clear that this value will be
eventually removed, and replaced by the lifecycle state feature.

Hence, this commit marks the value as deprecated. Moreover, we make
sure it will not be propagated to the resources (where it isn't
supported anyway):

1. effectiveStatus property will never contain this value;
2. and the default source of the administrativeStatus outbound mapping
("Magic" computed status) will never contain it.

This resolves MID-9026.