Remove setting delta old values by mappings
The ProjectionMappingSetEvaluator is not the right place to store
estimated old values for the deltas produced. The reason is that the
values are based on what mapping thinks was there, which may or may not
correspond to the reality. Lens context does the same, and it should
know better.

This should resolve MID-9102.