Check authorizations in some midpoint.* methods
When using methods countAccounts, getObjectsInConflictOnPropertyValue,
and isUniquePropertyValue we originally did not check authorizations
of the currently logged-in user. This is now fixed. Also, remaining
methods that avoid authorization checks were marked as such in the API
docs.

Note that this is an incompatible change, so users are advised to
check their configurations.

This resolves MID-6241.