Improve CompositeObjectDefinition performance
There was a performance-related bug introduced in midPoint 4.5
(a5ba3d72ad5600ac090e77d3345e6363c6ec7b8a), manifesting itself when
auxiliary object class definitions were used, and when there were many
attributes. (This is typical for LDAP.) The list of attributes got
re-computed unnecessarily often, so some simple operations like
searching for all accounts on a resource might take considerable time.

This commit introduces caching of the computed list, along with making
sure that all resource schemas are frozen just after being parsed.
(This is a prerequisite for the caching.)

The performance of the attribute lookup could be improved further, but
now it should be comparable to the one in 4.4.

This should resolve MID-9134.