Refactor role analysis detection and migration logic
Since this version, we support the application of detected patterns even in
the case of overlapping access rights. This option enables the creation of
business roles from the point of view hierarchy connections. If
the detected pattern is part of an already migrated business role, it will
be removed from the list of detected patterns.

We also display information about running tasks in session and clustering tables. Specifically, it is about role analysis clustering
and migration task (NOTE: if is the migration task in process, it is not possible to migrate to a new business role.).

This version also removes shortcomings such as duplications, bugs and documentation in the code.