Make archetypeRef selector consider assignments
The 'archetypeRef' selector traditionally matched the (effective)
'archetypeRef' object property. However, this does not work well
e.g. with the current GUI implementation for adding new objects,
where only the assignments are present. So this commit changes that.
See also the reasoning in the ArchetypeRefClause code.

(Note that this change involves only matching existing values,
not creating the filters.)

Related to MID-9268.