Introduce PolyString attributes in the repository
This commit brings in storing normalizable shadow attributes, i.e.,
those that have a matching/normalization rule defined, as PolyStrings.
Previously, these were stored as plain strings in their normalized form.
That was OK for searching, but is no longer suitable for full attribute
caching.

Major changes:

- Added NormalizationAwareResourceAttributeDefinition that supports
handling of these attributes. Also, the structure of classes related
to attribute definitions was changed: RawResourceAttributeDefinition
is not an interface, but a concrete class instead. The delegation
to rawDefinition in ResourceAttributeDefinition is no longer based on
the generic delegator (to increase clarity).
RawResourceObjectClassDefinition is gone.

- Furthered differentiation of shadow objects in provisioning-impl:
RawRepoShadow, RepoShadow, ResourceObject, ExistingResourceObject,
in order to know exactly the data structure we work with. The same
was done for the tests. Added convenience methods for retrieving
AbstractShadow instances in ProvisioningService. Goal: code clarity,
contract enforcement (also via increased consistency checking).

- Restructured processing of acquired shadows in provisioning
(classification, shadow update, combining shadow + resource object)
to ShadowPostProcessor. Goal: removing duplicate code.

Minor and/or unrelated changes:

- Fixed listing all dead shadows in lookupShadowByIndexedPrimaryIdValue
(this led to e.g. problems described in MID-9328).

- Renamed RepoShadowFinder back to ShadowFinder (to facilitate
backporting of fixes from 4.9 to 4.8 and earlier).

- The findAttributeDefinition method is now "<T> T" instead of "<?>".

- Consistence checks are now turned on for provisioning-impl tests.

- QShadowMapping: not retrieving index-only attributes on modifications.
This is a temporary change that is to be discussed.

- Adaptation to changes induced by
0213195e58a5c37826fd630d613c7dba57df9e5d in prism.

Related to MID-2119 (shadow caching).

Work in progress. Some changes are experimental/temporary.
Many tests currently fail.