Make provisioning work on generic repo
The change from string to polystring storage for normalized attributes
broke things for the generic repo. This commit tries to fix that:

1. "Polystring norm" queries for extension items and attributes were
not implemented correctly. Now they should be. Not in particularly nice
way, but probably good enough.

2. As part of the above, some parts of query interpreter were tidied
a bit. In particular, the earlier attempt of parameterizing
JpaDataNodeDefinition was reverted, as there was too many type safety
warnings that I was not able to resolve. Also, QueryInterpreterTest was
brought back to life, with some parts turned off (Hibernate 6).

3. A minor bug was fixed in the provisioning-impl.

Work in progress. The tests up to (including) provisioning-impl should
now be OK.

Related to MID-2119 (shadow caching).