Fix delayed-deletion of already disabled shadows
If the disabled shadow was to be delayed-deleted, for example,
when transitioning focus from active to suspended to archived
state with disable-instead-of-delete and delayed-delete features
set, the mechanism did not work, because the disable reason was
not updated from #mapped to #deprovision on the suspended->archived
focus transition.

To fix this, the behavior of "disableTimestamp" and "disableReason"
in the shadow activation container had to be changed. Originally,
these properties were updated only if there was an actual change in
the administrative status from something to DISABLED. Since this
commit, both of these properties are updated even if the administrative
status is already DISABLED (and the reason changes).

The only real code changes are in processActivationMetadata(..) method
in ActivationProcessor. All others are tests, docs, and cosmetics.

Related to MID-9220.