Support new associations
Associations are now stored in <associations> container structured
into individual multivalued items (like ri:group, ri:privilege, etc).
The item type is ShadowAssociationValueType.

Originally, they were in <association> container that had multiple
values of ShadowAssociationType, each of which carried the association
name. There was no item-like structure, only a plain list of values.

The current structure is very similar to the structure of attributes.

Related changes, stemming mainly from the efforts to simplify the code:

1. UCF is now expected to provide objects with the most current
definitions, i.e., with those that were originally called "refined"
ones. The reason is performance and code simplicity - avoiding
re-application of definitions. UCF is a regular part of midPoint,
and as such, it should have the most current definitions available.

2. UCF attribute conversion was improved: in the case of failure,
it returns all attributes that could be converted. Originally,
it returned only those that were converted before the first failure.

3. UcfChange has stronger requirements: (a) Primary identifier is now
required there. This simplifies the implementation. It also means that
asynchronous changes that fail the conversion from UcfChangeType bean
are simply discarded. There was nothing reasonable to be done with them,
anyway. (b) Also, the correct definitions must be applied to both object
and delta.

4. ShadowCaretaker is gone! The part related to applying the definitions
was moved into ShadowDefinitionApplicator (schema), with the perspective
of integrating it into standard applyDefinition mechanism.

5. Other code simplifications, e.g., in "resource objects" package,
and some parts of the projector. In particular, the recompute() method
in lens/focus/projection context was finally removed. It had been
practically "no-op" for some years now.

Work in progress. Main things to resolve:

1. Make sure that the UCF always has the most current schema definition.
2. Fix stray normalizations in ResourceAttribute (an artefact from
previous PolyString-based repo shadow storage attempts).