Improve fine-grained REST authorizations
This is an improvement of da667ef7debff76ef5e496913b85634ec2ec0650:
Instead of checking REST action authorization in the method body
(which was wrong), we determine the respective method, with the
corresponding action URI right in MidPointGuiAuthorizationEvaluator.
So the full REST-method-level authorization can be done there.

This is the correct solution. So, the hacks introduced previously
are rolled-back in this commit.