Separate assoc. simulation from other processing
This is an attempt to separate the definition of association simulation
(processed by the resource objects layer in provisioning) from the rest
of the processing (shadows layer, model, and so on).

The idea is that in the near future, connectors will provide native
support for associations. Hence, we will most probably define their
handling without the need of defining how they are simulated using
matching attributes (like ri:dn<->ri:member, ri:memberOf<->ri:dn, ...).

This is a highly experimental work. Nevertheless, it is going to the
master branch to avoid diverge too much. The legacy approach still
works, after all.

Main ideas:

1. The description how associations are simulated using attributes is
now part of the new "associations" capability.

2. There's a new multivalued "associationType" item in "schemaHandling".
Its purpose is to refine association classes (as provided by connector
or simulated) into association types, just like object classes are
refined into object types. It is very experimental, though.

3. There are three related definitions:

- ShadowAssociationTypeDefinition
- ShadowAssociationClassSimulationDefinition
- ShadowAssociationDefinition

See their respective docs. The first one is perhaps to be reconsidered.

Other changes:

1. The parsing process of complete resource schema was improved.
The code is cleaner. Errors are reported more comprehensively.
(E.g., delineation filters are parsed at parse time, not at runtime
as they were before.)

2. There are minor code improvements here and there. E.g.,
the entitlement converter code was streamlined a bit again.

3. Automatic caching of association binding attributes (the "value"
side) was disabled. It is recommended to mark them as secondary
identifiers.