Add "forced classification" option
Sometimes, we need to see the effect of changing classification rules
even on objects that are already classified. Originally, this was not
possible: if a shadow is classified, it will not be reclassified, not
even in simulation. Hence, this commit adds the "forced classification"
option that will trigger classification even on already-classified
shadows. It can be used both for the simulation and for the real
execution. However, for the time being it is reserved for internal use
only. It will be used by a dedicated "reclassify" task.

Related to MID-9514.