Roll back use of model-level resolver in tests
Introduced in 554eb0f3846cb99793e51ded5180a61f5aa5d5b8, the
default use of model-level object resolver (used mainly for
shadows) was not a great idea. A number of tests rely on
the "no fetch" resolution.

Hence, now it's rolled back. If needed, it has to be turned
on explicitly, like this:

assertUserAfter(userOid)
------> .withObjectResolver(createSimpleModelObjectResolver())
        .singleLink()
        .resolveTarget()
        .associations()
        .assertSize(1)
        ...

This should fix failing tests.