Fix caching shadow prism definition in ROCD/ROTD
In some cases, the ShadowType prism CTD was created before the object
type or class schema was completely parsed; in particular, before the
association definitions were added. This occurred e.g. when delineation
filter was present. The result was the inability to read the
associations.

This is now fixed by invalidating the cached definition appropriately.

This should resolve MID-9535.