Start implementing new format for associations
On 2024-04-09 we agreed a new format for the definition of associations.
The idea is to have more abstract representation visible to users,
while the technical form lies beneath.

Here is the preliminary implementation. Still a PoC. No outbounds yet.
Severe limitations for inbounds, e.g. correlation on targetRef does not
work yet. AD scenario also not supported (because of <ref>.</ref>).