Improve the associations definitions
The "association type" is now what it should be (according to the
conclusions on last office days): a logical connection between
a subject and 0-N objects.

Its definition is now more symmetrical and streamlined:
the complex value processing (including correlation and synchronization)
was moved into <valueProcessing> item available for all associations.

Work in progress: still not the final form, but getting closer.