Fix associationTargetSearch with multiple target types
This is a cherry-pick of ec5e25cfbbddb2c7de8dd2b5b1560a6fa9eed4d7,
adapted to the "new" associations. As part of it, the
ShadowAssociationDefinition#createTargetObjectsFilter() method
was improved - as much as currently possible.

Work in progress: there are some limitations remaining.

Related to MID-9561 and MID-9565.