Fix fetching incompatible associations
When having two account types, the first with an association defined
(account/IA), and second one without one (account/IB); and only ex-post
("after retrieval") classification using <classificationCondition>,
then searching for account/IA fails on a shadow S of type account/IB,
if it has a value for that association.

The reason is that the search for associations on S is based on
the definition of IA (so the association value is found), then comes
the classification that determines the type to be IB, and then
the association value becomes illegal.

This commit fixes that by removing such illegal association values.

This resolves MID-9591. However, the recommended way is to avoid
<classificationCondition> and to use delineation <filter> instead.
That way, all accounts found will be compatible with the follow-on
processing.