Improve authorization for filter items
In order to evaluate a filter, one has to be authorized to access
items (and their values) used for filter evaluation. The support
for this feature was present but a bit incomplete. "Deny"
authorizations were not taken into account, and authorizations
for unrelated types (required e.g. by the referencedBy filter)
were ignored.

This commit partially fixes that: "deny" authorizations are now
supported in the same way as "allow" ones, and some filter items
are checked, at least at a rudimentary level. To be improved later.

(Also adding forgotten TestExpressionProfiles to test suite.)

Related to MID-9638 and MID-9670.