Improve search item autz evaluation
To address MID-9638/MID-9670 (leaking data via referencedBy filter),
the handling of items allowed for search operations was changed.

It is now evaluated not only for the type we are searching for
(like RoleType), but for all types whose items are to be used
for the search (like UserType for a filter like "give me RoleType
referencedBy UserType via assignment/targetRef").

Limitations:

The checks are "yes/no" style only, based on the presence or absence
of authorizations against specified type and item(s), with appropriate
action URIs (read, search, and the new searchBy). No detailed checking
for the values is done. E.g. if the search for UserType:name is allowed
even for potentially a single user object (via an authorization clause
that can provide any number of matching objects, even zero), then the
"name" item can be used for any search concerning UserType or even
FocusType objects.

Effects on existing deployments:

1. Some queries allowed previously may now fail because of missing
item-searching authorizations. As a quick fix, new (experimental,
temporary) "searchBy" authorization is available to give search access
to these items without providing any additional access to data values.

2. Some queries denied previously may now be allowed. This should be
quite rare, but possible. (See TestSecurityAdvanced.test151 for an
example.) It can happen if the original authorization was not applied
because of some specific limitations (like roleRelation with no
explicit role information), and hence the item/exceptItem part of it
was skipped. This is no longer the case.