Support modern value metadata
This commit cuts off the legacy metadata for objects, assignments, and
credentials, and replaces it with "modern" value metadata.

Work in progress:

1. Model and workflows should be migrated. The main limitation is that
the provenance metadata for assignments are not fully supported yet.

2. Certifications were not touched yet. In GUI there were only minor
changes; the majority needs to be done there.

3. Data migration aspect was skipped for now. Most probably this will
have to be done during the upgrade, as the metadata handling is fairly
complex even now (because of multi-valued-ness of the data).