Sqale: Providee item definitions to parsers in repository.
In the case of repository - definition of parsed object is known beforehand
so this patch takes advantage of that and configures parsers
with item definition.

This decreases lookups into full schema registry.

Definitions are not cached in repository code itself, but cached and
retrieved from schema registry state, so in case of changing schemas
during runtime old definitions are purged with schema state.