sqale: Introduced m_object_view (view of all objects including shadows)
The way how filter generation works and use-cases in MidPoint shows
that we need to access shadows as objects sometimes and features
depends on it.

Introduced m_object_view view which is union of m_object and m_shadow.
This union has bit overhead in some queries, but also if it can be
proved that type is shadow (or not shadow) it does not perform
search in other part.