Support new associations (in progress)
This is the final iteration of the support for new associations.
The major trait is the distinction between higher-level concept
of associations (being more restricted, and therefore simpler to use),
and their lower-level implementation based on reference attributes
- highly flexible, powerful, open to future extension.

Work in progress. Tests up to provisioning-impl pass.
Caching of associations is not implemented yet, though.