Add preliminary reference attributes caching
ShadowDeltaComputerAbsolute and ShadowObjectComputer were updated to
cache reference attributes (ShadowDeltaComputerRelative was not - yet).

The post-processing of embedded shadows in references was moved from
ShadowedObjectConstruction to ShadowPostProcessor, so that OIDs are
available when shadows are updated (during post processing). Doing that
at the shadowed object construction time was too late.

Work in progress. Not all code paths are covered yet.