Improve association inbound/outbound definitions
In order to support various inbound/outbound mappings related settings
(like lifecycle state, strength, authoritativeness, and so on), and
to support multiple inbound/outbound mappings per association type,
we decided to include these in separate "inbound" and "outbound" items.

PRELIMINARY SOLUTION.

Differences to agreed-upon version:

1. For simplicity, the embedded attribute/objectRef mappings are not
named "mapping", but - as it was before - "inbound" and "outbound".
This was to avoid the need of cloning ResourceAttributeDefinitionType
just for this small change. It will be done later.

2. The "focus", "correlation" and "synchronization" sections have no
meaning outside the "inbound" mapping definition, as they are specific
for given inbound mapping. So, they were moved into it; and instead
of "focus", we now have (more traditional) "target" config item.