Clean up the associations schema a bit
Here we removed strange inbound/objectRef/inbound and
outbound/objectRef/outbound combinations of elements, and replace
them with the envisioned inbound/objectRef/mapping and
outbound/objectRef/mapping ones - with the benefits of making
the schema more direct (avoiding meaningless configurations at the
level of XSD), with the cost of duplication of parts of it.

Also removed some leftovers ("REMOVE ME") from the last commit.